worker/requirements.txt
celery
redis
flower
python-dotenv
